Implement Schedules API and integrate current API to client (#31)
* Implement Schedule entity, controller, service and all other related classes

* Add user role handling in UserPatcher for updating user details

* Add unique constraints for email and phone in User model; include schedules relationship

* Add schedule retrieval method in UserService to fetch user schedules

* Refactor TableService to use Integer type for exists and notExists methods

* Add endpoint to retrieve user schedule by ID

* Add ScheduleStatus enum and integrate status handling in Schedule model

* Add methods to ScheduleRepository and ScheduleService for date-based schedule retrieval

* Add endpoints for retrieving and adding user schedules by date

* Remove unused imports and whitespace in User and UserService classes

* Rename employee query hooks to user and update data types for consistency

* Add custom hook for querying user schedules by date

* Add custom hook for querying user schedules with error handling

* Refactor EmployeesList component to use UserData type and update role display logic

* Add user role and status enums, update UserData interface, and enhance Schedule interface

* Add utility functions for parsing and formatting dates and timestamps

* Update reservation modals to use createdFor instead of user for customer name

* Refactor EmployeeInfo component to use UserData type, update API calls to use users endpoint, and enhance role and status handling

* Add EmployeeSchedule component for managing employee schedules with date and time inputs

* Add EmployeeCalendar component for displaying and managing employee schedules

* Add RangeCalendarField component for integrating a range calendar with react-hook-form

* Enhance SelectField component to support required fields and improve value handling

* Refactor InputField component to enhance validation rules handling and support custom validation functions

* Enhance NumberField component to support min and max validation rules for numeric inputs

* Optimize TimeField component by memoizing parsed time calculations for improved performance

* Refactor Employee status handling to use UserStatus and integrate EmployeeSchedule component in EmployeeDetails

* Integrate EmployeeSchedule component and update user query in EmployeeDetails

* Update UserAvatar component to display email instead of username in description

* Add NotFound component and handle error state in EmployeeDetails

* Add methods to UserRepository for role-based user retrieval

* Add support for retrieving registered users by ID in UserController and UserService

* Add support for retrieving guest users in UserService and update UserController

* Update ReservationPatcher to handle null table IDs when patching reservations

* Remove cascade option from reservations mapping in Table model

* Update EditReservationModal to handle empty table ID values

* Remove redundant required attributes from SelectField component

* Add cascade option to reservations mapping in Table model

* Make schedule status field required in EmployeeSchedule component

* Add error handling to RangeCalendarField component

* Set default value for schedules in EmployeeSchedule component

* Rename user fields to createdBy and createdFor in CreateReservationModal

* Refactor ConfirmReservationModal to use react-query for reservation confirmation

* Refactor CancelReservationModal to use react-query for reservation cancellation

* Handle undefined values in SelectField component to prevent errors

* Refactor CreateReservationModal to simplify user input fields and update form validation mode

* Add custom hook useQueryUserByEmail for fetching user data by email

* Add label prop to InputField in NumberField component

* Handle null values for status and userRole in UserController before saving

* Refactor User model to handle null values for status and userRole, setting defaults when necessary

* Refactor CreateReservationModal to use react-query for reservation submission and simplify loading state management

* Update CreateReservationModal to conditionally include table ID in reservation data

* Change entity id types to Long

* Remove columnDefinition annotation from createdFor and createdBy, change id type to Long

* Improve formatting for clarity, change id type to Long

* Unset application.properties values and write comment to enforce editting of the configuration

+++ server/src/main/java/com/kalyvianakis/estiator/io/controller/ReservationController.java
@@ -38,7 +38,7 @@ public class ReservationController {
   }
 
   @GetMapping("/{id}")
-  public ResponseEntity<?> get(@PathVariable Integer id) throws ResourceNotFoundException {
+  public ResponseEntity<?> get(@PathVariable Long id) throws ResourceNotFoundException {
       return ResponseEntity.ok().body(reservationService.get(id));
   }
 
@@ -48,7 +48,7 @@ public class ReservationController {
   }
 
   @DeleteMapping("/{id}")
-  public ResponseEntity<?> delete(@PathVariable Integer id) throws ResourceNotFoundException {
+  public ResponseEntity<?> delete(@PathVariable Long id) throws ResourceNotFoundException {
       if (reservationService.notExists(id)) {
         throw new ResourceNotFoundException("Reservation not found for ID: " + id);
       }
@@ -58,7 +58,7 @@ public class ReservationController {
   }
 
   @PatchMapping("/{id}")
-  public ResponseEntity<?> patch(@PathVariable Integer id, @RequestBody Reservation reservation) throws ResourceNotFoundException {
+  public ResponseEntity<?> patch(@PathVariable Long id, @RequestBody Reservation reservation) throws ResourceNotFoundException {
       Reservation current = reservationService.get(id);
 
       reservationPatcher.patch(current, reservation);
