+++ server/src/main/java/com/kalyvianakis/estiator/io/repository/ReservationRepository.java
@@ -6,4 +6,4 @@ import org.springframework.stereotype.Repository;
 import com.kalyvianakis.estiator.io.model.Reservation;
 
 @Repository
-public interface ReservationRepository extends JpaRepository<Reservation, Integer> {}
+public interface ReservationRepository extends JpaRepository<Reservation, Long> {}
